requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.56.0
datasets==3.6.0
dill==0.3.8
docker==7.1.0
fastapi==0.115.12
filelock==3.13.1
Flask==3.1.1
fonttools==4.58.1
frozenlist==1.6.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lightning==2.5.1.post0
lightning-utilities==0.14.3
Mako==1.3.10
Markdown==3.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mlflow==2.22.1
mlflow-skinny==2.22.1
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
opentelemetry-api==1.34.0
opentelemetry-sdk==1.34.0
opentelemetry-semantic-conventions==0.55b0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
protobuf==6.31.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchmetrics==1.7.2
torchvision==0.22.1+cu128
tqdm==4.67.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
